# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tour/places/search": {
      "get": {
        "description": "예: \"성산일출봉\", \"흑돼지\". KorService2 `searchKeyword2`.\n\n응답 `id` = contentId. 후보 추가 시 `tourContentId`로 넘기거나 상세의 `placeId`를 쓰세요.",
        "operationId": "TourController_search",
        "parameters": [
          {
            "name": "keyword",
            "required": true,
            "in": "query",
            "description": "검색 키워드",
            "schema": {
              "example": "성산일출봉",
              "type": "string"
            }
          },
          {
            "name": "areaCode",
            "required": false,
            "in": "query",
            "description": "지역코드 (예: 39=제주)",
            "schema": {
              "example": "39",
              "type": "string"
            }
          },
          {
            "name": "sigunguCode",
            "required": false,
            "in": "query",
            "description": "시군구코드",
            "schema": {
              "example": "4",
              "type": "string"
            }
          },
          {
            "name": "contentTypeId",
            "required": false,
            "in": "query",
            "description": "12관광지 · 14문화 · 15축제 · 25코스 · 28레포츠 · 32숙박 · 38쇼핑 · 39음식",
            "schema": {
              "example": 12,
              "type": "number",
              "enum": [
                12,
                14,
                15,
                25,
                28,
                32,
                38,
                39
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "검색 결과 페이지",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourPlacePageDto"
                }
              }
            }
          }
        },
        "summary": "키워드로 관광지 검색",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/places/nearby": {
      "get": {
        "description": "GPS(mapX=경도, mapY=위도) 주변. 거리순 + `distanceMeters`. KorService2 `locationBasedList2`.",
        "operationId": "TourController_nearby",
        "parameters": [
          {
            "name": "mapX",
            "required": true,
            "in": "query",
            "description": "경도 (mapX)",
            "schema": {
              "example": 126.942,
              "type": "number"
            }
          },
          {
            "name": "mapY",
            "required": true,
            "in": "query",
            "description": "위도 (mapY)",
            "schema": {
              "example": 33.458,
              "type": "number"
            }
          },
          {
            "name": "radius",
            "required": false,
            "in": "query",
            "description": "반경(m), 최대 20000",
            "schema": {
              "default": 2000,
              "example": 2000,
              "type": "number"
            }
          },
          {
            "name": "contentTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 12,
              "type": "number",
              "enum": [
                12,
                14,
                15,
                25,
                28,
                32,
                38,
                39
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "주변 관광지 (distanceMeters 포함)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourPlacePageDto"
                }
              }
            }
          }
        },
        "summary": "내 위치 주변 관광지",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/places/sync": {
      "get": {
        "description": "캐시·DB 배치 갱신용. 일반 탐색은 `/tour/places` · `/tour/places/search`를 쓰세요.",
        "operationId": "TourController_sync",
        "parameters": [
          {
            "name": "areaCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "39",
              "type": "string"
            }
          },
          {
            "name": "sigunguCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "4",
              "type": "string"
            }
          },
          {
            "name": "contentTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 12,
              "type": "number",
              "enum": [
                12,
                14,
                15,
                25,
                28,
                32,
                38,
                39
              ]
            }
          },
          {
            "name": "modifiedtime",
            "required": false,
            "in": "query",
            "description": "이 시각 이후 수정분만 (YYYYMMDDHHMMSS)",
            "schema": {
              "example": "20260101000000",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 100,
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourSyncPageDto"
                }
              }
            }
          }
        },
        "summary": "[배치용] 동기화 목록",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/places/{contentId}/pet": {
      "get": {
        "description": "펫 동반 안내. 없으면 `null`. KorService2 `detailPetTour2`.",
        "operationId": "TourController_pet",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "1019041",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "펫 정보. 미등록이면 null",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourPetInfoDto"
                }
              }
            }
          }
        },
        "summary": "반려동물 동반 가능 여부",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/places/{contentId}/similar": {
      "get": {
        "description": "1순위: 티맵 연계 방문 데이터 (`TarRlteTarService1`).\n\n실패 시: 주변 같은 유형 관광지(KorService2)로 **자동 fallback**.\n\n`meta.source` = `RELATED_API` | `NEARBY_FALLBACK`.",
        "operationId": "TourController_similar",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "description": "예: 경복궁",
            "schema": {
              "example": "126508",
              "type": "string"
            }
          },
          {
            "name": "contentTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 12,
              "type": "number",
              "enum": [
                12,
                14,
                15,
                25,
                28,
                32,
                38,
                39
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourSimilarPageDto"
                }
              }
            }
          }
        },
        "summary": "비슷한 장소 (연관 관광지)",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/places/{contentId}/congestion": {
      "get": {
        "description": "`TatsCnctrRateService`. `cnctrRate` 0~100 (100=역대 최대 수준).\n\n`busy: true` = rate ≥ 60 → \"이번 주말 혼잡 예정\" 배지 권장.\n\nareaCd/signguCd 미지정 시 KorService2 상세의 areacode/sigungucode 사용 (코드 체계가 다르면 쿼리로 특화 코드를 직접 넘기세요).",
        "operationId": "TourController_congestion",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "126508",
              "type": "string"
            }
          },
          {
            "name": "areaCd",
            "required": false,
            "in": "query",
            "description": "특화 API areaCd. 미지정 시 KorService2 areacode 사용",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "signguCd",
            "required": false,
            "in": "query",
            "description": "특화 API signguCd. 미지정 시 KorService2 sigungucode 사용",
            "schema": {
              "example": "1110",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourCongestionDto"
                }
              }
            }
          }
        },
        "summary": "혼잡·집중률 예측 (향후 30일)",
        "tags": [
          "관광 탐색 · TourAPI",
          "관광 인사이트 · 빅데이터"
        ]
      }
    },
    "/api/v1/tour/places/{contentId}": {
      "get": {
        "description": "detailCommon+Intro+Info+Image 병렬 합침. Mongo `placeId`·`areaCode`·`sigunguCode` 포함.",
        "operationId": "TourController_detail",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "schema": {
              "example": "126508",
              "type": "string"
            }
          },
          {
            "name": "contentTypeId",
            "required": false,
            "in": "query",
            "schema": {
              "example": 12,
              "type": "number",
              "enum": [
                12,
                14,
                15,
                25,
                28,
                32,
                38,
                39
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourPlaceDetailDto"
                }
              }
            }
          }
        },
        "summary": "관광지 상세 (소개·이미지·부가정보 한 번에)",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/places": {
      "get": {
        "description": "areaCode / sigunguCode / contentTypeId. KorService2 `areaBasedList2`.",
        "operationId": "TourController_list",
        "parameters": [
          {
            "name": "areaCode",
            "required": false,
            "in": "query",
            "description": "지역코드 (areaCode)",
            "schema": {
              "example": "39",
              "type": "string"
            }
          },
          {
            "name": "sigunguCode",
            "required": false,
            "in": "query",
            "description": "시군구코드",
            "schema": {
              "example": "4",
              "type": "string"
            }
          },
          {
            "name": "contentTypeId",
            "required": false,
            "in": "query",
            "description": "12관광지 · 14문화 · 15축제 · 25코스 · 28레포츠 · 32숙박 · 38쇼핑 · 39음식",
            "schema": {
              "example": 12,
              "type": "number",
              "enum": [
                12,
                14,
                15,
                25,
                28,
                32,
                38,
                39
              ]
            }
          },
          {
            "name": "arrange",
            "required": false,
            "in": "query",
            "description": "A:제목 C:수정일 O:이미지+제목 Q:이미지+수정일",
            "schema": {
              "example": "O",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourPlacePageDto"
                }
              }
            }
          }
        },
        "summary": "지역별 관광지 목록",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/festivals": {
      "get": {
        "description": "eventStartDate(YYYYMMDD)부터. 미지정 시 오늘.",
        "operationId": "TourController_festivals",
        "parameters": [
          {
            "name": "eventStartDate",
            "required": false,
            "in": "query",
            "description": "행사 시작일 YYYYMMDD (미지정 시 오늘)",
            "schema": {
              "example": "20260101",
              "type": "string"
            }
          },
          {
            "name": "eventEndDate",
            "required": false,
            "in": "query",
            "description": "행사 종료일 YYYYMMDD",
            "schema": {
              "example": "20261231",
              "type": "string"
            }
          },
          {
            "name": "areaCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "39",
              "type": "string"
            }
          },
          {
            "name": "sigunguCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "4",
              "type": "string"
            }
          },
          {
            "name": "arrange",
            "required": false,
            "in": "query",
            "schema": {
              "example": "O",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourFestivalPageDto"
                }
              }
            }
          }
        },
        "summary": "행사 · 축제 목록",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/stays": {
      "get": {
        "description": "contentTypeId 32 숙박. 여행방 숙소 후보용.",
        "operationId": "TourController_stays",
        "parameters": [
          {
            "name": "areaCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "39",
              "type": "string"
            }
          },
          {
            "name": "sigunguCode",
            "required": false,
            "in": "query",
            "schema": {
              "example": "4",
              "type": "string"
            }
          },
          {
            "name": "arrange",
            "required": false,
            "in": "query",
            "schema": {
              "example": "O",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourPlacePageDto"
                }
              }
            }
          }
        },
        "summary": "숙박 목록",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/regions/{areaCd}/highlights": {
      "get": {
        "description": "`LocgoHubTarService1`. 지자체에서 가장 많이 연결되는 중심 스팟 (hubRank).\n\ndestinations seed / 지역 핵심 추천에 사용.\n\n특화 areaCd·signguCd 사용 (KorService2 코드와 다를 수 있음).",
        "operationId": "TourController_regionHighlights",
        "parameters": [
          {
            "name": "areaCd",
            "required": true,
            "in": "path",
            "description": "지역 코드 (예: 1=서울)",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "signguCd",
            "required": true,
            "in": "query",
            "description": "시군구 코드 (특화 API 체계). KorService2 sigungucode와 다를 수 있음",
            "schema": {
              "example": "1110",
              "type": "string"
            }
          },
          {
            "name": "baseYm",
            "required": false,
            "in": "query",
            "description": "기준 연월 YYYYMM (미지정 시 직전 월)",
            "schema": {
              "example": "202504",
              "type": "string"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourHubPageDto"
                }
              }
            }
          }
        },
        "summary": "시군구 중심 관광지 (허브)",
        "tags": [
          "관광 탐색 · TourAPI",
          "관광 인사이트 · 빅데이터"
        ]
      }
    },
    "/api/v1/tour/analytics/visitors": {
      "get": {
        "description": "`DataLabService`. KT(내국인)+SKT(외국인).\n\n- `level=metro` → 광역 (`metcoRegnVisitrDDList`)\n- `level=local` → 기초 (`locgoRegnVisitrDDList`)\n\n`visitorType`: domestic / other / foreign (코드 1·2·3 또는 K·E).\n월 단위 조회 권장. 광역·기초는 집계 기준이 달라 단순 합산 불가.",
        "operationId": "TourController_visitors",
        "parameters": [
          {
            "name": "level",
            "required": true,
            "in": "query",
            "description": "metro=광역(metco) · local=기초(locgo)",
            "schema": {
              "example": "metro",
              "type": "string",
              "enum": [
                "metro",
                "local"
              ]
            }
          },
          {
            "name": "startYmd",
            "required": true,
            "in": "query",
            "description": "시작일 YYYYMMDD",
            "schema": {
              "example": "20250401",
              "type": "string"
            }
          },
          {
            "name": "endYmd",
            "required": true,
            "in": "query",
            "description": "종료일 YYYYMMDD",
            "schema": {
              "example": "20250407",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 100,
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TourVisitorsPageDto"
                }
              }
            }
          }
        },
        "summary": "지역별 방문자수 (빅데이터)",
        "tags": [
          "관광 탐색 · TourAPI",
          "관광 인사이트 · 빅데이터"
        ]
      }
    },
    "/api/v1/tour/meta/ldong-codes": {
      "get": {
        "description": "미지정 시 시도 목록. `lDongRegnCd` 주면 시군구.",
        "operationId": "TourController_ldongCodes",
        "parameters": [
          {
            "name": "lDongRegnCd",
            "required": false,
            "in": "query",
            "description": "법정동 시도코드 (미지정 시 시도 목록)",
            "schema": {
              "example": "11",
              "type": "string"
            }
          },
          {
            "name": "lDongSignguCd",
            "required": false,
            "in": "query",
            "description": "법정동 시군구코드",
            "schema": {
              "example": "110",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 100,
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourCodeItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "[필터용] 법정동(시도·시군구) 코드",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
    "/api/v1/tour/meta/category-codes": {
      "get": {
        "description": "카테고리 필터. lclsSystm1/2로 하위 분류.",
        "operationId": "TourController_categoryCodes",
        "parameters": [
          {
            "name": "lclsSystm1",
            "required": false,
            "in": "query",
            "description": "분류체계 대분류 코드",
            "schema": {
              "example": "AC",
              "type": "string"
            }
          },
          {
            "name": "lclsSystm2",
            "required": false,
            "in": "query",
            "description": "분류체계 중분류 코드",
            "schema": {
              "example": "AC01",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "default": 100,
              "example": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TourCodeItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "[필터용] 분류체계 코드",
        "tags": [
          "관광 탐색 · TourAPI"
        ]
      }
    },
...
    "/api/v1/places/{placeId}/similar": {
      "get": {
        "description": "우리 DB에 저장된 장소의 태그/카테고리로 비슷한 곳을 찾습니다.\n\nTourAPI 관광지의 \"비슷한 장소\"는 `/tour/places/{contentId}/similar`를 쓰세요.",
        "operationId": "PlacesController_similar",
        "parameters": [
          {
            "name": "placeId",
            "required": true,
            "in": "path",
            "description": "Mongo placeId (ObjectId)",
            "schema": {
              "example": "665abc123def456789012345",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommonPlaceDto"
                  }
                }
              }
            }
          }
        },
        "summary": "유사 장소 (DB 태그·카테고리 기준)",
        "tags": [
          "장소 · Kakao/DB"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tourmate API",
    "description": "TripMatch / Tourmate backend REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "TourPlacePageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourPlaceCardDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/TourPageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "TourSyncPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourSyncCardDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/TourPageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "TourSimilarPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourRelatedPlaceDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/TourSimilarMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "TourFestivalPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TourFestivalCardDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/TourPageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}